# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: MedExpertMatch API
  description: |
    AI-powered medical expert recommendation system API.
    
    This API provides endpoints for:
    - Medical agent operations (matching, analysis, routing)
    - Synthetic data generation
    - Medical case and doctor management
    
    **Medical Disclaimer:**
    This API is for research and educational purposes only. It is NOT certified for clinical use 
    and should NOT be used for diagnostic decisions without human-in-the-loop verification.
  version: 1.0.0
  contact:
    name: MedExpertMatch API Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.medexpertmatch.com
    description: Production server
...
paths:
  /api/v1/evidence/verify-pubmed:
    get:
      tags:
        - Evidence
      summary: Verify PubMed tool
      description: |
        Verifies the PubMed tool by running a search and returning article count and sample.
        Use this to confirm PubMed is reachable from the app server and parsing works.
        NCBI E-utilities are called with tool/email and User-Agent for identification.
      operationId: verifyPubMed
      parameters:
        - name: query
          in: query
          required: false
          description: Search query string
          schema:
            type: string
            default: "GERD"
            example: "GERD"
        - name: maxResults
          in: query
          required: false
          description: Maximum number of articles to fetch (max 10)
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 3
            example: 3
      responses:
        '200':
          description: Verification result (ok or 0 articles)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyPubMedResponse'
        '500':
          description: PubMed request failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyPubMedResponse'
...
  /api/v1/synthetic-data/generate:
    post:
      tags:
        - Synthetic Data
      summary: Generate synthetic data
      description: |
        Generates synthetic data based on size parameter.
        Uses FHIR R5 compliant resources for data generation.
      operationId: generateSyntheticData
      parameters:
        - name: size
          in: query
          required: false
          description: Data size (tiny, micro, mini, compact, small, standard, medium, large, huge)
          schema:
            type: string
            enum: [ tiny, micro, mini, compact, small, standard, medium, large, huge ]
            default: tiny
        - name: clear
          in: query
          required: false
          description: Whether to clear existing data first
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Synthetic data generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticDataResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/synthetic-data/generate/doctors:
    post:
      tags:
        - Synthetic Data
      summary: Generate doctors only
      description: Generates synthetic doctor data.
      operationId: generateDoctors
      parameters:
        - name: count
          in: query
          required: false
          description: Number of doctors to generate
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 100
      responses:
        '200':
          description: Doctors generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticDataResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/synthetic-data/generate/cases:
    post:
      tags:
        - Synthetic Data
      summary: Generate medical cases only
      description: Generates synthetic medical case data.
      operationId: generateMedicalCases
      parameters:
        - name: count
          in: query
          required: false
          description: Number of cases to generate
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 200
      responses:
        '200':
          description: Medical cases generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticDataResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/synthetic-data/clear:
    post:
      tags:
        - Synthetic Data
      summary: Clear all synthetic data
      description: Clears all synthetic data from the database.
      operationId: clearSyntheticData
      responses:
        '200':
          description: Synthetic data cleared successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticDataResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    VerifyPubMedResponse:
      type: object
      required:
        - status
        - query
        - articleCount
        - message
      properties:
        status:
          type: string
          enum: [ ok, error ]
          description: Verification status
        query:
          type: string
          description: Search query used
        maxResults:
          type: integer
          description: Max results requested
        articleCount:
          type: integer
          description: Number of articles returned by PubMed
        message:
          type: string
          description: Human-readable message
        sample:
          type: array
          description: Sample articles (pmid, title) when articleCount > 0
          items:
            type: object
            properties:
              pmid:
                type: string
              title:
                type: string
...
    SyntheticDataResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum: [ success, error ]
          description: Operation status
          example: "success"
        message:
          type: string
          description: Status message
          example: "Synthetic data generated successfully"
        size:
          type: string
          description: Data size used (for generate endpoint)
          example: "medium"
        count:
          type: integer
          description: Number of items generated (for specific generation endpoints)
          example: 100
        cleared:
          type: boolean
          description: Whether data was cleared (for generate endpoint)
          example: false
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum: [ error ]
          example: "error"
        message:
          type: string
          description: Error message
          example: "Medical case not found: 696d4041ee50c1cfdb2e27ae"
        error:
          type: string
          description: Error type or code
          example: "NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-18T23:30:00Z"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if required)
security:
  - ApiKeyAuth: [ ]